requirements.txt
Django==4.1.7
djongo==1.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
requests==2.28.2
pytz==2023.3.post1
beautifulsoup4==4.12.2
selenium==4.12.0
pymongo==3.12.3
django-crontab==0.7.1